requirements/prod.txt
# Everything needed in production

# Flask
Flask==1.1.1
Werkzeug==1.0.0
click>=7.0

# Database
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.13
psycopg2-binary==2.8.4

# Redis
flask-redis==0.4.0

# Migrations
Flask-Migrate==2.5.2

# Task
celery==4.4.2

# Job
apscheduler==3.6.3
Flask-APScheduler==1.11.0

# JSON
jsonschema==3.2.0

# Forms
# Flask-WTF==0.14.3
# WTForms==2.2.1

# Deployment
gevent==1.4.0
gunicorn>=19.9.0
supervisor==4.1.0

# Flask Static Digest
# Flask-Static-Digest==0.1.3

# Auth
Flask-Login==0.5.0
Flask-Bcrypt==0.7.1
pyjwt==1.7.1

# Caching
Flask-Caching>=1.7.2

# Debug toolbar
Flask-DebugToolbar==0.10.1

# Environment variable parsing
environs==7.2.0
